moved some code away from app.ts
